requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.10.0
black==21.9b0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Django==3.2.8
django-cors-headers==3.10.0
django-environ==0.8.1
djangorestframework==3.12.4
flake8==4.0.1
idna==3.3
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2==2.9.1
py==1.10.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.3
regex==2021.10.8
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
